+++ kite-core/src/test/kotlin/com/tang/kite/session/SelectTest.kt
@@ -165,15 +165,6 @@ class SelectTest : BaseDataTest() {
     }
 
     @Test
-    fun paginateRequest() {
-        val session = sqlSessionFactory.openSession()
-        val accountMapper = session.getMapper(AccountMapper::class)
-        val page = accountMapper.paginate(request)
-        session.close()
-        assertNotEquals(0, page.total)
-    }
-
-    @Test
     fun selectOneToOneById() {
         val session = sqlSessionFactory.openSession()
         val accountMapper = session.getMapper(AccountOneToOneMapper::class)